@@ -3,20 +3,25 @@
   "description": "React Native Fireproof Example",
   "version": "0.0.1",
   "private": true,
+  "type": "module",
   "scripts": {
     "android": "react-native run-android",
     "ios": "react-native run-ios",
     "lint": "eslint .",
-    "pods": "bundle install && bundle exec pod install --project-directory=ios/",
+    "pods": "bundle install && bundle exec pod install --project-directory=ios/ && rm -rf .xcode.env.local",
     "start": "react-native start"
   },
   "dependencies": {
+    "@fireproof/react-native": "workspace:^",
     "@react-native-community/checkbox": "^0.5.17",
+    "@react-native-community/cli-platform-android": "14.0.0",
+    "@react-native/gradle-plugin": "^0.74.87",
     "react": "18.2.0",
     "react-native": "0.73.7",
     "react-native-fast-encoder": "^0.1.12",
+    "react-native-fs": "^2.20.0",
     "react-native-mmkv": "<3",
-    "react-native-quick-crypto": "margelo/react-native-quick-crypto",
+    "react-native-quick-crypto": "0.7.3",
     "use-fireproof": "file:../../fireproof/dist/use-fireproof/use-fireproof-0.0.0-smoke.tgz"
   },
   "devDependencies": {
@@ -24,18 +29,25 @@
     "@babel/plugin-transform-class-static-block": "^7.24.7",
     "@babel/preset-env": "^7.20.0",
     "@babel/runtime": "^7.20.0",
-    "@react-native/babel-preset": "0.73.21",
-    "@react-native/eslint-config": "0.73.2",
-    "@react-native/metro-config": "0.73.5",
-    "@react-native/typescript-config": "0.73.1",
+    "@eslint/compat": "^1.1.1",
+    "@eslint/js": "10.0.0",
+    "@react-native/babel-preset": "0.74.87",
+    "@react-native/eslint-config": "0.74.87",
+    "@react-native/eslint-plugin": "^0.74.87",
+    "@react-native/metro-config": "0.74.87",
+    "@react-native/typescript-config": "0.74.87",
     "@types/react": "^18.2.6",
     "@types/react-test-renderer": "^18.0.0",
+    "@typescript-eslint/eslint-plugin": "^8.0.1",
+    "@typescript-eslint/parser": "^8.0.1",
     "babel-jest": "^29.6.3",
-    "eslint": "^8.19.0",
+    "eslint": "9.9.0",
+    "eslint-plugin-react-native": "^4.1.0",
     "jest": "^29.6.3",
-    "prettier": "2.8.8",
+    "prettier": "3.3.3",
     "react-test-renderer": "18.2.0",
-    "typescript": "5.0.4"
+    "typescript": "5.5.4",
+    "typescript-eslint": "^8.0.1"
   },
   "engines": {
     "node": ">=18"
