@@ -1,18 +1,17 @@
-import React, { useState } from "react";
-import { Button, StyleSheet, Text, TextInput, View } from "react-native";
+import { useState } from 'react';
+import { Button, StyleSheet, Text, TextInput, View } from 'react-native';
 import Checkbox from '@react-native-community/checkbox';
-import { useFireproof } from "../../src/index";
-import { Todo, TodoFromAllDocs } from "./TodoList";
-// import { Doc } from "@fireproof/core";
+import { useFireproof } from '@fireproof/react-native';
+import type { Todo, TodoFromAllDocs } from './TodoList';
 
 type TodoItemProps = {
-  item: TodoFromAllDocs;
-}
-const TodoItem = ({item}: TodoItemProps) => {
+  readonly item: TodoFromAllDocs;
+};
+function TodoItem({ item }: TodoItemProps) {
   if (!item) {
     return null;
   }
-  const { database: db, useDocument } = useFireproof('TodoDB', {public: true});
+  const { database: db, useDocument } = useFireproof('TodoDB');
   const [todo, setTodo, saveTodo] = useDocument<Todo>(() => item.value);
   const [editMode, setEditMode] = useState<boolean>(false);
   const textStyle = todo.completed ? styles.completed : styles.notCompleted;
@@ -21,35 +20,33 @@ const TodoItem = ({item}: TodoItemProps) => {
     <View style={styles.itemRow}>
       <Checkbox
         value={todo.completed}
-        onValueChange={async (completed) => {
-          await saveTodo({...todo, completed});
+        onValueChange={async completed => {
+          await saveTodo({ ...todo, completed });
           setTodo();
         }}
         style={styles.checkbox}
       />
-      {
-        editMode ? (
-          <TextInput
-            style={textStyle}
-            value={todo.text}
-            onChangeText={(text) => setTodo({...todo, text})}
-            onSubmitEditing={() => {
-              saveTodo();
-              setEditMode(false)
-            }}
-          />
-        ) : (
-          <Text style={textStyle} onPress={() => setEditMode(true)}>
-            {todo.text}
-          </Text>
-        )
-      }
+      {editMode ? (
+        <TextInput
+          style={textStyle}
+          value={todo.text}
+          onChangeText={text => setTodo({ ...todo, text })}
+          onSubmitEditing={() => {
+            saveTodo();
+            setEditMode(false);
+          }}
+        />
+      ) : (
+        <Text style={textStyle} onPress={() => setEditMode(true)}>
+          {todo.text}
+        </Text>
+      )}
       <Button
         title="X"
         color="red"
         onPress={async () => {
           const del = await db.del(todo._id || '');
-          console.log({del});
+          console.log({ del });
         }}
       />
     </View>
